fix: Renter animation with `replay` when enemy re-enters stagger state
This is a dirty fix, I could image that a porper more general solution
would be to use a flag like `just_changed` in a state machine and then
only update the `AnimationPlayer2D` with a `.start()` when you enter a
new state.

If the same state was entered (like the Stagger state) then you simply
replay that animation. I don't see any reason why this wouldn't work,
but will have to see.