feat: Add first heavy attack (axe kick)
This was pretty straightforward to add, so I am hopeful that this
framework will work for this game.

I am not sure how the combo breaks have to be handled. Right now you can
just "soft"-combo, meaning once any attack is over and you enter the
recover state you can go with any attack again, I am not sure if that's
so good.

Perhaps one solution is to track if the player messed up a combo (for
example inputting a heavy attack when you can't chain it with the
current attack), in which case the recover animation will play out till
the end? Dunno, could also feel really frustrating.

Anyways, this marks the very first pass of the prototype, yay!