feat: Rework input system and add `InputDevice`
We are now in control of which device is managing the player. This
wouldn't quite work for local multiplayer (in which case you actually
have it easier because you just link one device to one player), but I
don't think that matters to much. It's important to get the basic game
done ASAP.