fix: Set enemy stagger state using `new_state`
Fixes #1.

This is actually a pretty important design pattern that I must use. We
are not allowed to set the enemy state outside of the
`EnemyStateSystemSet` (same with Player).

Instead one must set the `new_state` property.

I don't like this too much, the alternative is to use Events, though in
this case you need to pretty much do the exact same thing, it doesn't
change much in the underlying design pattern. You still need to be very
much aware of the scheduling of your functions.