feat!: Finish parry system, move ParryState into player statemachine
ParryState was moved into player statemachine and Component aspect was
removed. I plan to only have that on the player anyways and it's way
easier this way.

This is actually kinda crucial. This really showed me that a single
struct for the statemachine has big ups, you can easily check certain
state conditions (like if you can run/attack) because all the necessary
information is in this struct.

Additionally, you can have your own flags whenever you want to set a
value. Of course the downside is that it isn't very modular at all, but
I feel like for this game that absolutely doesn't matter. As there are
only going to be players and enemies at most.