fix: Split logic of putting enemy in stance break state
The issue was that scheduling was ambigious and as a result we could get
into a scheduling where we never stance break.

I also decided to put the reset_just_changed into PreUpdate because it
should pretty much fire before anything else. We will see if that is the
right decision or not.