feat!: Replace `DudeState::Sliding` with `Attack::Slide`
The slide is now an attack, I decided this is a better way to handle
any state that has a hitbox (spawns an attack arc).

This may not be quite true for things that don't spawn an attack arc,
but then you can just spawn an invisible attack arc with just a hitbox,
so even then putting it into `Attack` is not a bad move.

It also works way better with `StaggerState` and other parts.