feat!: Add stagger recover animations and add fall stagger
Restructured the whole Stagger imlpementation.

I moved the Stagger struct into the statemachines and removed the
Component derive. This is better, but still not ideal, I have some
gripes about this implementation but it will do for now.

One possible improvement would be to use something similar to the
Jumping state, where we have `DudeState::Stagger(StaggerState)`, but I
don't quite know how that would work with the struct (storing
information on the enums).