feat: Implement flying attacks hitboxes
This isn't the most beautiful imlpementation but oh well. I do want to
point out that the hitbox doesn't move when jumping. This is by design!
I realized while playtesting that this works better, though it actually
was more work to get this to work than if I had the hitbox moving along.

The fact that the hitbox stays in position help to sell the illusion of
depths, without it the player hits enemies that he should not be able to
hit.