feat!: Add basic (and broken) pathfinding using A* (A-Star)
There are still a lot of issues with this, however it is practically
impossible for me to work on this let alone fix them. That is why I am
restructuring the whole process and try to simplify it as much as
possible.

It seems like this is about my limit for technical debt, at least for
now. About 1k+ lines in 5+ files, with a fairly well documented
algorithm that isn't terribly hard to understand. Oh well, I will take
it.