ci/cd: Remove wasm build from CD pipeline
Dropped wasm support as it is a pain in the ass and not a target for the
final game anyways. Perhaps I will make some kind of demo in the future
where I specifically build for wasm and then never update it, but I
definitely don't want to support wasm throughout the project.