feat: Stop player movement when attack has activated hitbox
This fixes the issue of the player moving when he really shouldn't.
Ideally you would want to define this using an animation curve, but this
will do for now (I searched for an animation curve but just couldn't
find any standalone tool for that, also integrating it wouldn't be super
easy).