fix: Skip level transition when position is `Vec2::ZERO`
This should fix #5. It's not a super clean solution but I doubt it would
ever be an issue in the actual game. It may be solved in the future
anyways, but it probably doesn't hurt to keep this as a Vec2 being ZERO
is very unlikely and even if the way the level transition is set up this
should not be an issue.