refactor: Replace state logic to reflect the pathfinding logic updates
This is the result of the decoupling of pathfinding/target logic and
state logic. We only care about the enemy in the state logic, which is
exactly what we want. There is no need for any pathfinding logic to be
there, this reduceses the queries and the overall logic of the states.

This is really important as the states are one of the hardest things to
maintain, so you kind of want to optimize for minimization of them.