restructure!: Remove redundant groups in hitbox and hurtbox
I had a logical error in my thinking when I was creating the structure
of the hitbox/hurtbox backend. I misunderstood the concept of
memberships and filters in the bitmasks.

The bitmasks are always going to lead to a collision because it's always
like this

.. / membership - filter
Hit  Hitbox       Hurtbox
Hurt Hurtbox      Hitbox

No matter what other types of groups we add to either, we will
**always** get collision between the two. This commit removes the
redundant groups as they don't matter and only lead to confusion.

Instead of using the groups, use `Hitbox` to infer what kind of attack
was issued and by whom. I am still not completely sold on this as it's a
little convoluted but I think this would even work decently well for
enemy on enemy interaction so should be fine for this game.