fix: Trigger `LevelChanged` even in `PreUpdate` to make sure all events fire before level transitions
I can't believe this bug has been here this whole time... So what
happened was that the `LevelChanged` event got fired after some other
systems ran that were listening for it, then in the `PostUpdate` we
transition back to the normal game state and we can't listen for the
events anymore.

Really makes me wonder why we even need to run that whole thing in the
`LevelTransition` event in the first place. Wouldn't it be enough if we
just listened for the event? Ah well, past me probably had good reasons
to implement it this way.