restructure!: Rename `PlayerInput` -> `GamingInput`
The reason I renamed this resource is because I am planning to have
different input resources for different game states, for example I need
a way to get player input while in game over state, so I would add a
`MenuInput` that only runs when inside of a menu.

This could also be extended to have a global input for things like
entering debug mode or toggling fullscreen etc.

This idea sound pretty good right now, it would even allow the player to
customize the inputs more granularly (movement in game is independent
from movement in menu etc). Though we will see how it goes.