feat!: Support square tiles and diagonal tiles in the same level
The algorithm now merges them together and checks if the levels are all
valid (both in terms of pathfinding and in terms of graphics).

I honestly don't even know how necessary or useful this will be to the
artist, I do think that we want this, but it could very well be that
it's not required at all.